[RESTEASY-3424] Move the arquillian Jetty based tests to a single module in the testsuite.
https://issues.redhat.com/browse/RESTEASY-3424
