+++ .github/workflows/maven.yml
@@ -9,7 +9,6 @@ on:
       - '**'
     paths:
       - '**/pom.xml'
-      - 'arquillian/**'
       - 'providers/**'
       - 'resteasy-cdi/**'
       - 'resteasy-client/**'
