--- arquillian/.gitignore
@@ -1,1 +1,0 @@
-/as7-dist-async
