--- arquillian/RESTEASY-1056-jetty-bv11/.gitignore
@@ -1,1 +1,0 @@
-/target/
