--- arquillian/RESTEASY-1056-jetty-bv11/src/main/java/org/jboss/resteasy/resteasy1056/TestApplication.java
@@ -1,23 +1,0 @@
-package org.jboss.resteasy.resteasy1056;
-
-import java.util.HashSet;
-import java.util.Set;
-
-import jakarta.ws.rs.core.Application;
-import jakarta.ws.rs.ext.Provider;
-
-/**
- *
- * @author <a href="ron.sigal@jboss.com">Ron Sigal</a>
- * @version $Revision: 1.1 $
- *
- *          Copyright June 7, 2014
- */
-@Provider
-public class TestApplication extends Application {
-    public Set<Class<?>> getClasses() {
-        HashSet<Class<?>> classes = new HashSet<Class<?>>();
-        classes.add(TestResource.class);
-        return classes;
-    }
-}
