--- arquillian/RESTEASY-1056-jetty-bv11/src/test/resources/arquillian.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<arquillian xmlns="http://jboss.org/schema/arquillian" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://jboss.org/schema/arquillian
-        http://jboss.org/schema/arquillian/arquillian_1_0.xsd">
-
-    <!-- Force the use of the Servlet 5.0 protocol with all containers, as it is the most mature -->
-    <defaultProtocol type="Servlet 5.0" />
-
-    <engine>
-        <property name="deploymentExportPath">target/deployments</property>
-    </engine>
-
-    <container qualifier="jetty" default="true">
-        <configuration>
-            <property name="bindHttpPort">0</property>
-            <property name="javaVmArguments">${additionalJvmArgs}</property>
-            <!--property name="javaVmArguments">${additionalJvmArgs} -Xdebug -Xrunjdwp:transport=dt_socket,address=8585,server=y, suspend=y</property-->
-        </configuration>
-    </container>
-</arquillian>
