--- arquillian/RESTEASY-1630-jetty-resteasy-servlet-initializer/src/main/java/org/jboss/resteasy/resteasy1630/TestApplication.java
@@ -1,16 +1,0 @@
-package org.jboss.resteasy.resteasy1630;
-
-import java.util.HashSet;
-import java.util.Set;
-
-import jakarta.ws.rs.core.Application;
-import jakarta.ws.rs.ext.Provider;
-
-@Provider
-public class TestApplication extends Application {
-    public Set<Class<?>> getClasses() {
-        HashSet<Class<?>> classes = new HashSet<Class<?>>();
-        classes.add(TestResource.class);
-        return classes;
-    }
-}
