--- arquillian/RESTEASY-1630-jetty-resteasy-servlet-initializer/src/main/java/org/jboss/resteasy/resteasy1630/TestApplicationPath.java
@@ -1,8 +1,0 @@
-package org.jboss.resteasy.resteasy1630;
-
-import jakarta.ws.rs.ApplicationPath;
-import jakarta.ws.rs.core.Application;
-
-@ApplicationPath("/two")
-public class TestApplicationPath extends Application {
-}
