--- arquillian/RESTEASY-736-jetty/.gitignore
@@ -1,2 +1,0 @@
-/target
-.DS_Store
