--- arquillian/RESTEASY-736-jetty/src/test/resources/web.xml
@@ -1,25 +1,0 @@
-<web-app version="3.0" xmlns="http://java.sun.com/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd">
-
-    <display-name>RESTEASY-736</display-name>
-
-    <servlet>
-        <servlet-name>Resteasy</servlet-name>
-
-        <!--org.jboss.resteasy.plugins.server.servlet.HttpServletDispatcher-->
-        <servlet-class>
-            org.jboss.resteasy.plugins.server.servlet.HttpServlet30Dispatcher
-        </servlet-class>
-        <init-param>
-            <param-name>jakarta.ws.rs.Application</param-name>
-            <param-value>org.jboss.resteasy.resteasy736.TestApplication</param-value>
-        </init-param>
-        <async-supported>true</async-supported>
-    </servlet>
-
-    <servlet-mapping>
-        <servlet-name>Resteasy</servlet-name>
-        <url-pattern>/*</url-pattern>
-    </servlet-mapping>
-
-</web-app>
