--- arquillian/pom.xml
@@ -1,59 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <parent>
-        <groupId>org.jboss.resteasy</groupId>
-        <artifactId>resteasy-jaxrs-all</artifactId>
-        <version>6.2.6.Final-SNAPSHOT</version>
-        <relativePath>../pom.xml</relativePath>
-    </parent>
-    <name>RESTEasy Misc Arquillian-based tests</name>
-    <description/>
-    <modelVersion>4.0.0</modelVersion>
-    <packaging>pom</packaging>
-
-    <modules>
-        <module>RESTEASY-736-jetty</module>
-        <module>RESTEASY-1056-jetty-bv11</module>
-        <module>RESTEASY-1630-jetty-resteasy-servlet-initializer</module>
-    </modules>
-    
-    <artifactId>resteasy-misc-arquillian-tests</artifactId>
-
-    <properties>
-        <!-- Version properties -->
-        <version.arquillian.jetty>1.0.0.Final</version.arquillian.jetty>
-    </properties>
-
-    <dependencyManagement>
-        <dependencies>
-            <dependency>
-                <groupId>org.jboss.arquillian</groupId>
-                <artifactId>arquillian-bom</artifactId>
-                <version>${dep.arquillian-bom.version}</version>
-                <scope>import</scope>
-                <type>pom</type>
-            </dependency>
-            <dependency>
-                <groupId>org.jboss.arquillian.container</groupId>
-                <artifactId>arquillian-jetty-embedded-9</artifactId>
-                <version>${version.arquillian.jetty}</version>
-            </dependency>
-        </dependencies>
-    </dependencyManagement>
-
-    <build>
-        <plugins>
-            <plugin>
-                <artifactId>maven-surefire-plugin</artifactId>
-                <configuration>
-                    <skip>true</skip>
-                    <systemPropertyVariables>
-                        <additionalJvmArgs>${additionalJvmArgs}</additionalJvmArgs>
-                    </systemPropertyVariables>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-
-</project>
-
