+++ testsuite/jetty-integration-tests/src/test/java/org/jboss/resteasy/test/resteasy1056/TestApplication.java
@@ -1,0 +1,23 @@
+package org.jboss.resteasy.test.resteasy1056;
+
+import java.util.HashSet;
+import java.util.Set;
+
+import jakarta.ws.rs.ApplicationPath;
+import jakarta.ws.rs.core.Application;
+
+/**
+ *
+ * @author <a href="ron.sigal@jboss.com">Ron Sigal</a>
+ * @version $Revision: 1.1 $
+ *
+ *          Copyright June 7, 2014
+ */
+@ApplicationPath("/")
+public class TestApplication extends Application {
+    public Set<Class<?>> getClasses() {
+        HashSet<Class<?>> classes = new HashSet<Class<?>>();
+        classes.add(TestResource.class);
+        return classes;
+    }
+}
