+++ testsuite/jetty-integration-tests/src/test/java/org/jboss/resteasy/test/resteasy1630/TestApplication.java
@@ -1,0 +1,16 @@
+package org.jboss.resteasy.test.resteasy1630;
+
+import java.util.HashSet;
+import java.util.Set;
+
+import jakarta.ws.rs.core.Application;
+import jakarta.ws.rs.ext.Provider;
+
+@Provider
+public class TestApplication extends Application {
+    public Set<Class<?>> getClasses() {
+        HashSet<Class<?>> classes = new HashSet<Class<?>>();
+        classes.add(TestResource.class);
+        return classes;
+    }
+}
