+++ testsuite/jetty-integration-tests/src/test/java/org/jboss/resteasy/test/resteasy1630/TestApplicationPath.java
@@ -1,0 +1,8 @@
+package org.jboss.resteasy.test.resteasy1630;
+
+import jakarta.ws.rs.ApplicationPath;
+import jakarta.ws.rs.core.Application;
+
+@ApplicationPath("/two")
+public class TestApplicationPath extends Application {
+}
