+++ testsuite/jetty-integration-tests/src/test/resources/arquillian.xml
